split up constants
