+++ creusot/src/backend.rs
@@ -3,3 +3,4 @@ pub(crate) mod logic;
 pub(crate) mod program;
 pub(crate) mod term;
 pub(crate) mod traits;
+pub(crate) mod constant;
